rename mapper state stuff
